+++ frontend/src/components/SelectContainer/Timer/Timer.styled.ts
